Trees are initialized with their tree types

See merge request acx677/artificial-intelligence-software-agents!10

+++ TreeModel/Model/Tree/ITreeLayer.cs
@@ -50,7 +50,7 @@ namespace TreeModel.Model.Tree
         /// <returns>
         /// The age
         /// </returns>
-        int GetAge(Position tree);
+        double GetAge(Position tree);
 
         /// <summary>
         /// Inspection of the State of a tree
